+++ Specs/Core/Matrix4Spec.js
@@ -545,6 +545,24 @@ defineSuite([
         expect(result).toEqual(expected);
     });
 
+    it('multiplyByPoint works without a result parameter', function() {
+        var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
+        var right = new Cartesian4(17, 18, 19);
+        var expected = new Cartesian4(114, 334, 554, 774);
+        var result = left.multiplyByPoint(right);
+        expect(result).toEqual(expected);
+    });
+
+    it('multiplyByPoint works with a result parameter', function() {
+        var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
+        var right = new Cartesian4(17, 18, 19);
+        var expected = new Cartesian4(114, 334, 554, 774);
+        var result = new Cartesian4();
+        var returnedResult = left.multiplyByPoint(right, result);
+        expect(returnedResult).toBe(result);
+        expect(result).toEqual(expected);
+    });
+
     it('multiplyByScalar works without a result parameter', function() {
         var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
         var right = 2;
@@ -1207,6 +1225,20 @@ defineSuite([
         var matrix = new Matrix4();
         expect(function() {
             Matrix4.multiplyByVector(matrix, undefined);
+        }).toThrow();
+    });
+
+    it('static multiplyByPoint throws with no matrix parameter', function() {
+        var cartesian = new Cartesian4();
+        expect(function() {
+            Matrix4.multiplyByPoint(undefined, cartesian);
+        }).toThrow();
+    });
+
+    it('static multiplyByPoint throws with no cartesian parameter', function() {
+        var matrix = new Matrix4();
+        expect(function() {
+            Matrix4.multiplyByPoint(matrix, undefined);
         }).toThrow();
     });