+++ Specs/Scene/CompositePrimitiveSpec.js
@@ -3,6 +3,7 @@ defineSuite([
          'Scene/CompositePrimitive',
          'Specs/createContext',
          'Specs/destroyContext',
+         'Specs/createCamera',
          'Specs/equals',
          'Specs/frameState',
          'Specs/pick',
@@ -21,6 +22,7 @@ defineSuite([
          CompositePrimitive,
          createContext,
          destroyContext,
+         createCamera,
          equals,
          frameState,
          pick,
@@ -54,18 +56,13 @@ defineSuite([
     beforeEach(function() {
         primitives = new CompositePrimitive();
 
-        camera = new Camera(context.getCanvas());
+        camera = createCamera(context);
         camera.position = new Cartesian3(1.02, 0.0, 0.0);
         camera.up = Cartesian3.UNIT_Z;
         camera.direction = camera.position.normalize().negate();
-        camera.frustum.near = 0.01;
-        camera.frustum.far = 10.0;
-        camera.frustum.fovy = CesiumMath.toRadians(60.0);
-        camera.frustum.aspectRatio = 1.0;
 
         us = context.getUniformState();
-        us.setView(camera.getViewMatrix());
-        us.setProjection(camera.frustum.getProjectionMatrix());
+        us.update(camera);
         us.setSunPosition(new Cartesian3(-2.0, 0.0, 0.0));
     });